+++ ui/main/game/galactic_war/cards/gwaio_start_nomad.js
@@ -179,7 +179,7 @@ define([
               {
                 file: unit,
                 path: "command_caps",
-                op: "push",
+                op: "replace",
                 value: ["ORDER_Move", "ORDER_Patrol", "ORDER_Assist"],
               },
               {
@@ -236,6 +236,31 @@ define([
                 value: "ORDER_Use",
               }
             );
+          });
+
+          var offensiveStructures = [
+            "/pa/units/land/air_defense/air_defense.json",
+            "/pa/units/land/artillery_short/artillery_short.json",
+            "/pa/units/land/land_mine/land_mine.json",
+            "/pa/units/land/laser_defense_single/laser_defense_single.json",
+            "/pa/units/sea/torpedo_launcher/torpedo_launcher.json",
+            "/pa/units/land/air_defense_adv/air_defense_adv.json",
+            "/pa/units/land/laser_defense_adv/laser_defense_adv.json",
+            "/pa/units/land/laser_defense/laser_defense.json",
+            "/pa/units/land/tactical_missile_launcher/tactical_missile_launcher.json",
+            "/pa/units/orbital/ion_defense/ion_defense.json",
+            "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv.json",
+            "/pa/units/land/artillery_long/artillery_long.json",
+            "/pa/units/land/nuke_launcher/nuke_launcher.json",
+            "/pa/units/orbital/defense_satellite/defense_satellite.json",
+          ];
+          offensiveStructures.forEach(function (unit) {
+            mods.push({
+              file: unit,
+              path: "command_caps",
+              op: "push",
+              value: "ORDER_Attack",
+            });
           });
 
           inventory.addMods(mods);